manpages: prepare for new manpage format
This commit adds a upcoming manpage format to the AsciiDoc
backend. The new format changes are:

 * The synopsis is now a section with a dedicated style. This
 "synopsis" style allows to automatically format the keywords as
 monospaced and <placeholders> as italic.
 * the backticks are now used to format synopsis-like syntax in inline
 elements.

The parsing of synopsis is done with a new AsciiDoc extension that
makes use of the PEG parser parslet.

All the asciidoc manpages sources are processed with this
extension. It may upset the formatting for older manpages, making it
not consistent across a page, but this will be a mild side effect, as
this was not really consistent before.

Signed-off-by: Jean-Noël Avila <jn.avila@free.fr>